requirements.txt
flwr==1.7.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
